refactor: move filter-card to facet-filter
