+++ client-app/shared/catalog/components/index.ts
@@ -16,4 +16,4 @@ export { default as ProductsFiltersSidebar } from "./products-filters.vue";
 export { default as CategorySelector } from "./category-selector.vue";
 export { default as DiscountBadge } from "./discount-badge.vue";
 export { default as Vendor } from "./vendor.vue";
-export { default as FacetValues } from "./facet-values.vue";
+export { default as FacetFilter } from "./facet-filter.vue";
